added user & token authentication

+++ part4/bloglist/controllers/blogs.js
@@ -1,13 +1,17 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/Blog");
+const User = require("../models/user");
+const { userExtractor } = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", userExtractor, async (request, response) => {
   const body = request.body;
+
+  const user = await User.findById(request.user.id);
 
   if (!body.title || !body.url) {
     return response.status(400).json({ error: "content missing" });
@@ -18,13 +22,25 @@ blogsRouter.post("/", async (request, response) => {
     author: body.author,
     url: body.url,
     likes: body.likes ? body.likes : 0,
+    user: user.id,
   });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+  const user = request.user;
+  const blog = await Blog.findById(request.params.id);
+  if (blog.user.toString() !== user.id.toString()) {
+    return response
+      .status(401)
+      .json({ error: "token missing or invalid user" });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
